\name{ssgsea.GBM.classification-package}
\alias{ssgsea.GBM.classification-package}
\alias{ssgsea.GBM.classification}
\docType{package}
\title{
\packageTitle{ssgsea.GBM.classification}
}
\description{
\packageDescription{ssgsea.GBM.classification}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ssgsea.GBM.classification}
\packageIndices{ssgsea.GBM.classification}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ssgsea.GBM.classification}

Maintainer: \packageMaintainer{ssgsea.GBM.classification}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
